ai-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.0
google-generativeai==0.3.2
gitpython==3.1.40
pyyaml==6.0.1
requests==2.31.0
whisper==1.1.10
coqui-tts==0.20.6
pyaudio==0.2.11
speechrecognition==3.10.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
pyttsx3==2.90
